flake8 and move get_mpl_axes from figure.py to tools.py
